将 `damage` 与 `health` 更改为 double；添加了部分关键词；添加了新的 Bee, Place 类等

+++ include/Ants/AntRemover.h
@@ -9,7 +9,7 @@ private:
     static const int foodCost = 1;                    //!< 食物消耗（没错，Remover 工具也要消耗）
 
 public:
-    explicit AntRemover(int health = 0);
+    explicit AntRemover(double health = 0.0);
 };
 
 #endif // ANTREMOVER_H
